feat：support circuitbreaker by service and interface

+++ polaris-plugins/polaris-plugins-observability/stat-common/src/main/java/com/tencent/polaris/plugins/stat/common/model/StatInfoStatefulCollector.java
@@ -49,7 +49,7 @@ public class StatInfoStatefulCollector<T> extends AbstractSignatureStatInfoColle
                         metric = metricContainer.get(signature);
                         if (null == metric) {
                             StatStatefulMetric statMetric = new StatStatefulMetric(metricName, labels, signature);
-                            statMetric.setValue(strategy.initMetricValue(info));
+                            statMetric.setValue((long) strategy.initMetricValue(info));
                             LOG.debug("{} with signature {} init value is {}", strategy.getStrategyName(),
                                     signature,
                                     statMetric.getValue());
